Add more JUnit tests
