+++ data/duke.txt
@@ -1,3 +1,4 @@
 [T][ ] read
 [T][ ] write
 [E][ ] movie night (at: 18 Sep 2021)
+[D][ ] iP (by: 23:59)
