+++ src/test/java/duke/tasks/DeadlineTest.java
@@ -1,5 +1,6 @@
 package duke.tasks;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
@@ -11,5 +12,10 @@ public class DeadlineTest {
     public void invalidBy() {
         assertThrows(DukeException.class, () ->
                 new Deadline("task", "23 aug 2021"));
+    }
+
+    @Test
+    public void createDeadline() throws DukeException {
+        assertEquals("[D][ ] iP (by: 23:59)", new Deadline("iP", "23:59").toString());
     }
 }
